+++ common/src/main/java/top/vmctcn/vmtranslationupdate/modpack/VersionChecker.java
@@ -1,7 +1,6 @@
-package top.vmctcn.vmtranslationupdate.util;
+package top.vmctcn.vmtranslationupdate.modpack;
 
-import net.minecraft.entity.player.PlayerEntity;
-import net.minecraft.text.Text;
+import top.vmctcn.vmtranslationupdate.VMTranslationUpdate;
 
 import java.io.BufferedReader;
 import java.io.InputStreamReader;
@@ -9,10 +8,10 @@ import java.net.URI;
 import java.net.URLConnection;
 import java.nio.charset.StandardCharsets;
 
-public class VersionCheckUtil {
-    public static String getOnlineVersion(PlayerEntity player) {
+public class VersionChecker {
+    public static String getOnlineVersion() {
         try {
-            URI uri = URI.create(ModConfigUtil.getConfig().modPackTranslationUpdateCheckUrl);
+            URI uri = URI.create(ModpackInfoReader.getModpackInfo().getModpack().getTranslation().getUpdateCheckUrl());
             URLConnection connection = uri.toURL().openConnection();
 
             String userAgent = "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/109.0.5414.120 Safari/537.36 MCMod/VmUpdate";
@@ -23,7 +22,7 @@ public class VersionCheckUtil {
                 return reader.readLine();
             }
         } catch (Exception e) {
-            player.sendMessage(Text.translatable("vmtranslationupdate.message.error"));
+            VMTranslationUpdate.LOGGER.warn("Version check failed: ", e);
             return "";
         }
     }