+++ fabric/src/main/java/top/vmctcn/vmtranslationupdate/fabric/VMTranslationUpdateClientFabric.java
@@ -1,26 +1,17 @@
 package top.vmctcn.vmtranslationupdate.fabric;
 
 import net.fabricmc.api.ClientModInitializer;
-import net.fabricmc.fabric.api.client.event.lifecycle.v1.ClientTickEvents;
 import net.fabricmc.fabric.api.client.screen.v1.ScreenEvents;
 import top.vmctcn.vmtranslationupdate.ModEvents;
 import top.vmctcn.vmtranslationupdate.VMTranslationUpdate;
-import top.vmctcn.vmtranslationupdate.util.ScreenUtil;
-import top.vmctcn.vmtranslationupdate.util.ModConfigUtil;
 
 public class VMTranslationUpdateClientFabric implements ClientModInitializer {
     @Override
     public void onInitializeClient() {
         VMTranslationUpdate.init();
 
-        ClientTickEvents.END_CLIENT_TICK.register(client -> {
-            if (ModConfigUtil.getConfig().displayTips) {
-                ModEvents.clientTickEndEvent(client);
-            }
-        });
-
         ScreenEvents.AFTER_INIT.register((client, screen, scaledWidth, scaledHeight) -> {
-            ScreenUtil.screenAfterInitEvent(screen);
+            ModEvents.screenAfterInitEvent(screen);
         });
     }
 }