+++ neoforge/src/main/java/top/vmctcn/vmtranslationupdate/neoforge/ModPlatformImpl.java
@@ -1,9 +1,0 @@
-package top.vmctcn.vmtranslationupdate.neoforge;
-
-import net.neoforged.fml.ModList;
-
-public class ModPlatformImpl {
-    public static boolean isModLoaded(String modid) {
-        return ModList.get().isLoaded(modid);
-    }
-}